+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new `WorkspaceTabSelected` event
 - new `WorkspaceTabClicked` event
 - selection of panel initialization script 
+- role model for workspaces
 
 ### Changed